Add web-browsing control.
Add a WebView control to the document and add WebKit as a linked
framework. Change the sample TextField to display the URL from the
WebView and post a URL to the WebView. Set the constraints on the
TextField and WebView to preserve their relative spacing as their
window is resized (that is, they change size with the window).

Set the browser window’s size originally to that of the original
Macintosh’s screen. But then use the old width as the new height,
rounding it to the nearest multiple of 3 and using 4/3 of that height
as the width. Set the minimum size of the browser window with the same
formula, but a quarter of the area.