Add first attempt at local load & save support.
Set property list to indicate that the app can read any non-directory
file, write the same or web-archives, and acts as a non-originator of
those formats.

Add a file-export menu item.

Add private properties to the Document class to save a file’s
attributes on document-load so they can be used when the web-view is
later loaded from the document’s XIB. Use the web-view’s data as the
source during saving. Disable auto-saving. Touch-up XIB loading.