Improve opening links to new windows.
Add a delay between creating a browser window and loading the default
page (when a local file is not used). This adds a oppurtunity for the
loading to be cancelled, preventing wastes of time and work before a
new-window-with-link command sets its window’s initial link. (Without
this, the user can see the default page briefly before the desired
first link takes over.)