Add action to activate URL entry field.
Add menu command to highlight the text within a URL entry field, so a
new URL can be entered (and visited by the corresponding web-view). If
the loading bar is hidden, it will be made visible first. If there are
no browser windows, a new one will be created first.

Refactor to a new method the code to create a browser window (and
document) without an initial page visit.

Unexpectedly touch a previously untested web-view UI delegate method.