Add basic printing support.
Change the action of the menu command for printing so individual
controls won’t respond to it. Implement that action for the browser
window controller with a panel sheet for the settings.

Implement the action for an application-global Page Setup menu command
through a subclass of NSApplication. Adjust the project to use said
subclass for the application object.