Change how open-file success is determined.
Remove the success criterion of immediate successful creation of the
browser window that loads the file. Offload file opening to an
NSOperation subclass. Said subclass determines the success criterion as
each file finishes loading (successfully or not), then publishes the
merged result to the application object.

Enable browser window controller instances to notify other objects when
a resource has finished loading, successfully or not.