Add open-then-print menu command.
Add alternate menu command for Print that brings up an Open panel then,
if not cancelled, opens the selected files in new browser windows and
then, if not cancelled from the Print panels brought out next, prints
from those windows. Serve as a way to test the Print-Document Apple
event with minimal fuss.