Add storage of tracked visits between sessions.
Add two actions between the global web-history instance and its storage
file, one to load to the object and one to save to the file. Add user
preference to load the web-history store on launch and save it on
termination. Add a private preference storing a file bookmark to the
web-history store. Since bookmarks work only on existing files, use
various properties for generating the default URL for the store.