Add actions to revisit previous pages.
Add dynamically generated menu items to the menu for each visited day
in the global web-history instance’s database, each menu item having a
submenu for the individual visited pages. Add an action for the menu
items representing pages to have the corresponding page loaded in the
frontmost browser window (creating a window first, if needed).

Add a class facilitating updates to the tracked-visit menu items
whenever the global web-history instance changes. Add an instance of
this class to the application delegate and add a method to update the
menu to the delegate.