Refine handling of changed menu items.
Add (currently unused) code to handle a history menu item being
replaced. This would be called if a history menu item representing a
day is swapped out, and not if it’s mutated nor if its submenu or any
menu item of that submenu are mutated or removed/added/replaced.