Move search- or print-upon-loading code.
Move the code that triggers a string search or a print job upon a file
being opened from the file-opening class to the browser window
controller. This allows the search or print data to be part of the
Resume data when User Interface Preservation is implemented as part of
Automatic Termination.