Add User Interface Preservation (i.e. Resume).
Use the browser window controller class name as the Preservation
identifier for the browser window. Factor the generation of this name
out to a new method.

Set the browser window controller class as the Preservation manager for
browser windows. Add a property to the controller for indicating when
it’s under restore. Add methods and key strings to preserve and
recreate browser windows. Add and modify methods to reset the
restoration-flag as needed during extended state restoration.