Add "View Source" command and window.
Add menu command to view the source text of a loaded page, if the page’s
data type supports it. Add a new window XIB and controller class to
implement the command’s results.

Factor out the code for the application delegate to track new window
controllers, since there are now two kinds of user windows.